protoc-bundled-plugin: add protoc binaries for protobuf v3.3.0
The latest stable release of protobuf is
https://github.com/google/protobuf/releases/v3.3.0

This adds the binaries for that release and adds a script that
semi-automates this task.

This patch was created by running that script and manually
adding <artifact> entries for each binary to the POM.

The script was derived from the process used for PR 41 which
added binaries for v3.0.0.